Updated solution and refactored UserController to use service
Updated Visual Studio solution file to include a new project and updated the version. Added two new files to the project file for compilation. Refactored UserController to use a service for user management, which is injected via the constructor. Added new files defining an interface and a class for the user service, with the class implementing the interface using a static list of users.